am243x: DPL: Fix MisraC and SA issues
Fix all the MISRA-C violations for DPL No RTOS files
for Common, m4 and r5 files

Fixes: MCUSDK-10911

+++ source/kernel/nortos/dpl/r5/HwiP_armv7r_vim.h
@@ -105,7 +105,7 @@ static inline void  HWI_SECTION HwiP_setAsFIQ(uint32_t intNum, uint32_t isFIQ)
 
     if(isFIQ != 0U)
     {
-        *addr |= (0x1u << bitPos);
+        *addr |= ((uint32_t)0x1u << bitPos);
     }
     else
     {
@@ -135,7 +135,7 @@ static inline void  HWI_SECTION HwiP_setAsPulse(uint32_t intNum, uint32_t isPuls
 
     if(isPulse != 0U)
     {
-        *addr |= (0x1u << bitPos);
+        *addr |= ((uint32_t)0x1u << bitPos);
     }
     else
     {
